refactoring and fix minor bugs

+++ tests/test_PRTree.py
@@ -4,6 +4,7 @@ import numpy as np
 import os
 import time
 
+
 class TestPRTree(unittest.TestCase):
     def test_result(self):
         def has_intersect(x, y):
@@ -12,7 +13,7 @@ class TestPRTree(unittest.TestCase):
             return b1 and b2
 
         idx = np.arange(100)
-        x = np.random.rand(len(idx), 4).astype(np.float32)
+        x = np.random.rand(len(idx), 4)
         x[:, 1] += x[:, 0]
         x[:, 3] += x[:, 2]
 
@@ -21,7 +22,6 @@ class TestPRTree(unittest.TestCase):
         for i in range(len(idx)):
             tmp = [k for k in range(len(idx)) if has_intersect(x[i], x[k])]
             self.assertEqual(set(out[i]), set(tmp))
-
 
         prtree.save('tree.bin')
         time.sleep(.3)
@@ -40,11 +40,10 @@ class TestPRTree(unittest.TestCase):
             tmp = [k for k in range(len(idx)) if has_intersect(x[i], x[k])]
             self.assertEqual(set(out[i]), set(tmp))
 
-
         os.remove('tree.bin')
-        N= 100000
+        N = 100000
         idx = np.arange(N)
-        x = np.random.rand(N, 4).astype(np.float32)
+        x = np.random.rand(N, 4)
         x[:, 1] = x[:, 0] + x[:, 1] / np.sqrt(N) / 100
         x[:, 3] = x[:, 2] + x[:, 3] / np.sqrt(N) / 100
         prtree1 = PRTree(idx, x)
@@ -53,18 +52,19 @@ class TestPRTree(unittest.TestCase):
         for i in range(N//2, N):
             prtree2.insert(idx[i], x[i])
 
-        x = np.random.rand(N, 4).astype(np.float32)
+        x = np.random.rand(N, 4)
         x[:, 1] = x[:, 0] + x[:, 1] / np.sqrt(N) / 100
         x[:, 3] = x[:, 2] + x[:, 3] / np.sqrt(N) / 100
         for i in range(N):
-            self.assertEqual(set(prtree1.query(x[i])), set(prtree2.query(x[i])))
+            self.assertEqual(
+                set(prtree1.query(x[i])), set(prtree2.query(x[i])))
 
         for i in range(N-1):
             prtree1.erase(i)
             prtree2.erase(i)
 
+
 def suite():
     suite = unittest.TestSuite()
     suite.addTests(unittest.makeSuite(TestPRTree))
     return suite
-
